src/model_training/requirements.txt
boto3
numpy
pandas
mlflow
prefect
scikit-learn
hyperopt
python-dotenv
flask
